Refactor rendering to allow for subobjects
This adds the ability to add objects to your objects. The subobject of
an object is positioned relative to it's parent object.

Unfortunately to accomplish this we need to clone the subobject to
adjust it's X property without modifying it permanantly or having to
have a saveX property or something like that, but it should likely be
quickly discarded once it has been painted, which is when this takes
place.